Fix linter with robocop
