+++ exit_program.rb
@@ -1,0 +1,4 @@
+def exit_program
+    puts 'Thank you for using this app!'
+    exit
+end
