+++ rental_date.rb
@@ -1,0 +1,4 @@
+def rental_date
+    print 'Date (YYYY/MM/DD): '
+    gets.chomp
+end
